End of Line source code uploaded
A group effort with Greg Ward and Aaron Vos. A multiplayer game using
the Microsoft Kinect as a controller and WPF as the graphics foundation.
Two players can compete for survival in our take on the classic Tron
light bike scenario. The Microsoft Kinect SDK v1.8 and Kinect toolkit
may or may not have to be installed before running this application.